+++ src/energy_systems/general/bounded_supply.jl
@@ -6,7 +6,7 @@ component or other equipment unit that processes energy in a given medium. The c
 might still have a maximum power draw in a single time step, but can provide any fraction
 of this to connected components.
 """
-mutable struct BoundedSupply <: ControlledComponent
+mutable struct BoundedSupply <: Component
     uac::String
     controller::Controller
     sys_function::SystemFunction
@@ -15,16 +15,19 @@ mutable struct BoundedSupply <: ControlledComponent
     input_interfaces::InterfaceMap
     output_interfaces::InterfaceMap
 
-    max_power_profile::Profile
+    max_power_profile::Union{Profile,Nothing}
     temperature_profile::Union{Profile,Nothing}
     scaling_factor::Float64
 
     max_energy::Float64
     temperature::Temperature
+    static_power::Union{Nothing,Float64}
     static_temperature::Temperature
 
     function BoundedSupply(uac::String, config::Dict{String,Any})
-        max_power_profile = Profile(config["max_power_profile_file_path"])
+        max_power_profile = "max_power_profile_file_path" in keys(config) ?
+                            Profile(config["max_power_profile_file_path"]) :
+                            nothing
         temperature_profile = "temperature_profile_file_path" in keys(config) ?
                               Profile(config["temperature_profile_file_path"]) :
                               nothing
@@ -49,7 +52,8 @@ mutable struct BoundedSupply <: ControlledComponent
             config["scale"], # scaling_factor
             0.0, # max_energy
             nothing, # temperature
-            default(config, "static_temperature", nothing) # static_temperature
+            default(config, "static_power", nothing), # static_power
+            default(config, "static_temperature", nothing), # static_temperature
         )
     end
 end
@@ -75,16 +79,29 @@ function control(
     parameters::Dict{String,Any}
 )
     move_state(unit, components, parameters)
-    unit.max_energy = unit.scaling_factor * Profiles.work_at_time(unit.max_power_profile, parameters["time"])
+
+    if unit.static_power !== nothing
+        unit.max_energy = watt_to_wh(unit.static_power)
+    elseif unit.max_power_profile !== nothing
+        unit.max_energy = unit.scaling_factor * Profiles.work_at_time(
+            unit.max_power_profile, parameters["time"]
+        )
+    else
+        unit.max_energy = 0.0
+    end
     set_max_energy!(unit.output_interfaces[unit.medium], unit.max_energy)
 
     if unit.static_temperature !== nothing
         unit.temperature = unit.static_temperature
     elseif unit.temperature_profile !== nothing
-        unit.temperature = Profiles.value_at_time(unit.temperature_profile, parameters["time"])
+        unit.temperature = Profiles.value_at_time(
+            unit.temperature_profile, parameters["time"]
+        )
     end
-    unit.output_interfaces[unit.medium].temperature = highest_temperature(unit.temperature, unit.output_interfaces[unit.medium].temperature)
-
+    unit.output_interfaces[unit.medium].temperature = highest_temperature(
+        unit.temperature,
+        unit.output_interfaces[unit.medium].temperature
+    )
 end
 
 function process(unit::BoundedSupply, parameters::Dict{String,Any})