+++ src/energy_systems/storage/seasonal_thermal_storage.jl
@@ -4,7 +4,7 @@ Implementation of a seasonal thermal storage component.
 This is a simplified model, which mostly deals with amounts of energy and considers
 temperatures only for the available temperature as the tank is depleted.
 """
-mutable struct SeasonalThermalStorage <: ControlledComponent
+mutable struct SeasonalThermalStorage <: Component
     uac::String
     controller::Controller
     sys_function::SystemFunction
@@ -154,7 +154,7 @@ function load(unit::SeasonalThermalStorage, parameters::Dict{String,Any})
 end
 
 function output_values(unit::SeasonalThermalStorage)::Vector{String}
-    return ["IN", "OUT", "Load", "Capacity"]
+    return ["IN", "OUT", "Load", "Load%", "Capacity"]
 end
 
 function output_value(unit::SeasonalThermalStorage, key::OutputKey)::Float64
@@ -164,6 +164,8 @@ function output_value(unit::SeasonalThermalStorage, key::OutputKey)::Float64
         return calculate_energy_flow(unit.output_interfaces[key.medium])
     elseif key.value_key == "Load"
         return unit.load
+    elseif key.value_key == "Load%"
+        return 100 * unit.load / unit.capacity
     elseif key.value_key == "Capacity"
         return unit.capacity
     end