+++ test/energy_systems/gasboiler_demand_driven_with_bus.jl
@@ -35,7 +35,8 @@ function test_gasboiler_demand_driven_with_bus()
             "is_source" => true,
         ),
         "TST_GB_01" => Dict{String,Any}(
-            "type" => "GasBoiler",
+            "type" => "FuelBoiler",
+            "m_fuel_in" => "m_c_g_natgas",
             "control_refs" => ["TST_BUS_01"],
             "output_refs" => ["TST_BUS_01"],
             "strategy" => Dict{String,Any}(
@@ -103,7 +104,7 @@ function test_gasboiler_demand_driven_with_bus()
     EnergySystems.process(bus, simulation_parameters)
     EnergySystems.process(gasboiler, simulation_parameters)
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].balance ≈ 12000/4
-    @test gasboiler.input_interfaces[gasboiler.m_gas_in].balance ≈ -12000/4
+    @test gasboiler.input_interfaces[gasboiler.m_fuel_in].balance ≈ -12000/4
 
     EnergySystems.process(grid, simulation_parameters)
     @test grid.output_interfaces[grid.medium].balance ≈ 0
@@ -160,7 +161,7 @@ function test_gasboiler_demand_driven_with_bus()
     EnergySystems.process(bus, simulation_parameters)
     EnergySystems.process(gasboiler, simulation_parameters)
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].balance ≈ 12000/4
-    @test gasboiler.input_interfaces[gasboiler.m_gas_in].balance ≈ -12000/4
+    @test gasboiler.input_interfaces[gasboiler.m_fuel_in].balance ≈ -12000/4
 
     EnergySystems.process(grid, simulation_parameters)
     @test grid.output_interfaces[grid.medium].balance ≈ 0
@@ -217,7 +218,7 @@ function test_gasboiler_demand_driven_with_bus()
     EnergySystems.process(bus, simulation_parameters)
     EnergySystems.process(gasboiler, simulation_parameters)
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].balance ≈ 10000/4
-    @test gasboiler.input_interfaces[gasboiler.m_gas_in].balance ≈ -10000/4
+    @test gasboiler.input_interfaces[gasboiler.m_fuel_in].balance ≈ -10000/4
 
     EnergySystems.process(grid, simulation_parameters)
     @test grid.output_interfaces[grid.medium].balance ≈ 0
@@ -251,7 +252,8 @@ function test_gasboiler_demand_driven_without_bus()
             "is_source" => true,
         ),
         "TST_GB_01" => Dict{String,Any}(
-            "type" => "GasBoiler",
+            "type" => "FuelBoiler",
+            "m_fuel_in" => "m_c_g_natgas",
             "control_refs" => ["TST_DEM_01"],
             "output_refs" => ["TST_DEM_01"],
             "strategy" => Dict{String,Any}(
@@ -299,7 +301,7 @@ function test_gasboiler_demand_driven_without_bus()
     @test exchange.energy_potential ≈ 12000/4
     
     # no processing so far, balance is zero, energy_potential is non-zero
-    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_gas_in], gasboiler)
+    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_fuel_in], gasboiler)
     @test exchange.balance ≈ 0.0
     @test exchange.storage_potential ≈ 0.0
     @test exchange.energy_potential == -Inf
@@ -317,12 +319,12 @@ function test_gasboiler_demand_driven_without_bus()
     EnergySystems.process(gasboiler, simulation_parameters)
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].balance ≈ 0
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].sum_abs_change ≈ 2*12000/4
-    @test gasboiler.input_interfaces[gasboiler.m_gas_in].balance ≈ -12000/4
+    @test gasboiler.input_interfaces[gasboiler.m_fuel_in].balance ≈ -12000/4
     @test demand.input_interfaces[demand.medium].balance ≈ 0
     @test demand.input_interfaces[demand.medium].sum_abs_change ≈ 2*12000/4
 
     # demand and gasboiler were processed --> energy_potential should be zero, but not the balance
-    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_gas_in], gasboiler)
+    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_fuel_in], gasboiler)
     @test exchange.balance ≈ -12000/4
     @test exchange.storage_potential ≈ 0.0
     @test exchange.energy_potential ≈ 0.0
@@ -356,7 +358,7 @@ function test_gasboiler_demand_driven_without_bus()
     @test exchange.energy_potential ≈ 15000/4
     
     # no processing so far, balance is zero, energy_potential is non-zero
-    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_gas_in], gasboiler)
+    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_fuel_in], gasboiler)
     @test exchange.balance ≈ 0.0
     @test exchange.storage_potential ≈ 0.0
     @test exchange.energy_potential == -Inf
@@ -374,12 +376,12 @@ function test_gasboiler_demand_driven_without_bus()
     EnergySystems.process(gasboiler, simulation_parameters)
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].balance ≈ -15000/4 + 12000/4
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].sum_abs_change ≈ 12000/4 + 15000/4
-    @test gasboiler.input_interfaces[gasboiler.m_gas_in].balance ≈ -12000/4
+    @test gasboiler.input_interfaces[gasboiler.m_fuel_in].balance ≈ -12000/4
     @test demand.input_interfaces[demand.medium].balance ≈ -15000/4 + 12000/4   
     @test demand.input_interfaces[demand.medium].sum_abs_change ≈ 15000/4 + 12000/4
 
     # demand and gasboiler were processed --> energy_potential should be zero, but not the balance
-    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_gas_in], gasboiler)
+    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_fuel_in], gasboiler)
     @test exchange.balance ≈ -12000/4
     @test exchange.storage_potential ≈ 0.0
     @test exchange.energy_potential ≈ 0.0
@@ -413,7 +415,7 @@ function test_gasboiler_demand_driven_without_bus()
     @test exchange.energy_potential ≈ 10000/4
     
     # no processing so far, balance is zero, energy_potential is non-zero
-    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_gas_in], gasboiler)
+    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_fuel_in], gasboiler)
     @test exchange.balance ≈ 0.0
     @test exchange.storage_potential ≈ 0.0
     @test exchange.energy_potential == -Inf
@@ -431,12 +433,12 @@ function test_gasboiler_demand_driven_without_bus()
     EnergySystems.process(gasboiler, simulation_parameters)
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].balance ≈ 0
     @test gasboiler.output_interfaces[gasboiler.m_heat_out].sum_abs_change ≈ 2*10000/4 
-    @test gasboiler.input_interfaces[gasboiler.m_gas_in].balance ≈ -10000/4
+    @test gasboiler.input_interfaces[gasboiler.m_fuel_in].balance ≈ -10000/4
     @test demand.input_interfaces[demand.medium].balance ≈ 0
     @test demand.input_interfaces[demand.medium].sum_abs_change ≈ 2*10000/4 
 
     # demand and gasboiler were processed --> energy_potential should be zero, but not the balance
-    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_gas_in], gasboiler)
+    exchange = EnergySystems.balance_on(gasboiler.input_interfaces[gasboiler.m_fuel_in], gasboiler)
     @test exchange.balance ≈ -10000/4
     @test exchange.storage_potential ≈ 0.0
     @test exchange.energy_potential ≈ 0.0