+++ test/energy_systems/storage_loading_switch.jl
@@ -23,7 +23,8 @@ function test_primary_producer_can_load_storage()
             "is_source" => true,
         ),
         "TST_GBO_01" => Dict{String,Any}(
-            "type" => "GasBoiler",
+            "type" => "FuelBoiler",
+            "m_fuel_in" => "m_c_g_natgas",
             "control_refs" => ["TST_BFT_01"],
             "output_refs" => [
                 "TST_BUS_01"
@@ -36,7 +37,8 @@ function test_primary_producer_can_load_storage()
             "power" => 10000
         ),
         "TST_GBO_02" => Dict{String,Any}(
-            "type" => "GasBoiler",
+            "type" => "FuelBoiler",
+            "m_fuel_in" => "m_c_g_natgas",
             "control_refs" => ["TST_BFT_01"],
             "output_refs" => [
                 "TST_BUS_01"